requirements.txt
# Blogger API and auth
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# API server
uvicorn[standard]>=0.23.0
fastapi>=0.104.0
python-multipart>=0.0.6

# Draft script (frontmatter + markdown → HTML)
pyyaml>=6.0
markdown>=3.5
requests>=2.28.0
